[Task #1652] Remove TermAssignmentDao.java
